Change uint32_t -> size_t in *Enumerate functions
Follows https://github.com/webgpu-native/webgpu-headers/pull/142

Bug: dawn:160
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/75481
Commit-Queue: Austin Eng <enga@chromium.org>

+++ src/dawn_native/Adapter.cpp
@@ -89,7 +89,7 @@ namespace dawn_native {
         return mSupportedFeatures.IsEnabled(feature);
     }
 
-    uint32_t AdapterBase::APIEnumerateFeatures(wgpu::FeatureName* features) const {
+    size_t AdapterBase::APIEnumerateFeatures(wgpu::FeatureName* features) const {
         return mSupportedFeatures.EnumerateFeatures(features);
     }
