+++ src/dawn_native/Device.h
@@ -259,7 +259,7 @@ namespace dawn_native {
 
         bool APIGetLimits(SupportedLimits* limits) const;
         bool APIHasFeature(wgpu::FeatureName feature) const;
-        uint32_t APIEnumerateFeatures(wgpu::FeatureName* features) const;
+        size_t APIEnumerateFeatures(wgpu::FeatureName* features) const;
         void APIInjectError(wgpu::ErrorType type, const char* message);
         bool APITick();
