+++ src/dawn_wire/client/Device.h
@@ -67,7 +67,7 @@ namespace dawn_wire::client {
 
         bool GetLimits(WGPUSupportedLimits* limits) const;
         bool HasFeature(WGPUFeatureName feature) const;
-        uint32_t EnumerateFeatures(WGPUFeatureName* features) const;
+        size_t EnumerateFeatures(WGPUFeatureName* features) const;
         void SetLimits(const WGPUSupportedLimits* limits);
         void SetFeatures(const WGPUFeatureName* features, uint32_t featuresCount);
