+++ src/dawn_wire/client/LimitsAndFeatures.cpp
@@ -32,7 +32,7 @@ namespace dawn_wire::client {
         return mFeatures.count(feature) != 0;
     }
 
-    uint32_t LimitsAndFeatures::EnumerateFeatures(WGPUFeatureName* features) const {
+    size_t LimitsAndFeatures::EnumerateFeatures(WGPUFeatureName* features) const {
         if (features != nullptr) {
             for (WGPUFeatureName f : mFeatures) {
                 *features = f;
